+++ Konigsberg/LeetCode/Problem9.cs
@@ -1,0 +1,8 @@
+using Konigsberg.Libraries;
+
+namespace Konigsberg.LeetCode;
+
+public sealed class Problem9
+{
+    public bool IsPalindrome(int x) => x.IsPalindromic();
+}
